WIP: switch encoding unsigned type normalization to Uint* types
Using Java native types for unsigned values is wasteful especially
in case of uint32 and uint64. Furthermore it leads to uncertainty
between in16/uint8, int32/uint16 and int64/uint32.

This is a partial switch, where we retain java.lang types for
signed entities, but switch the unsigned ones.

JIRA: YANGTOOLS-1018

+++ yang/yang-data-impl/src/main/java/org/opendaylight/yangtools/yang/data/impl/codec/Uint16StringCodec.java
@@ -7,17 +7,18 @@
  */
 package org.opendaylight.yangtools.yang.data.impl.codec;
 
+import org.opendaylight.yangtools.yang.common.Uint16;
 import org.opendaylight.yangtools.yang.data.api.codec.Uint16Codec;
 import org.opendaylight.yangtools.yang.model.api.type.Uint16TypeDefinition;
 
-final class Uint16StringCodec extends AbstractIntegerStringCodec<Integer, Uint16TypeDefinition> implements
+final class Uint16StringCodec extends AbstractIntegerStringCodec<Uint16, Uint16TypeDefinition> implements
         Uint16Codec<String> {
     Uint16StringCodec(final Uint16TypeDefinition typeDef) {
-        super(typeDef, extractRange(typeDef), Integer.class);
+        super(typeDef, extractRange(typeDef), Uint16.class);
     }
 
     @Override
-    protected Integer deserialize(final String stringRepresentation, final int base) {
-        return Integer.valueOf(stringRepresentation, base);
+    protected Uint16 deserialize(final String stringRepresentation, final int base) {
+        return Uint16.valueOf(stringRepresentation, base);
     }
 }
